requirements.txt
mujoco>=3.0.0
mujoco_mjx>=3.0.0
brax>=0.10.0
playground
jax>=0.4.0
jaxlib>=0.4.0
flax>=0.7.0
optax>=0.1.0
numpy>=1.24.0
matplotlib>=3.6.0
mediapy>=1.1.0
ml_collections>=0.1.0
orbax-checkpoint>=0.2.0
etils>=1.3.0
IPython>=8.0.0
tqdm>=4.64.0
typer>=0.9.0
rich>=13.0.0
